Maxnet

Documentation for Maxnet.

Maxnet.maxnet

Maxnet.maxnet — Method

maxnet(
    presences, predictors, [features];
    regularization_multiplier, regularization_function,
    addsamplestobackground, weight_factor, backend, 
    kw...
)

Arguments

presences: A BitVector where presences are true and background samples are false

predictors: A Tables.jl-compatible table of predictors. Categorical predictors should be CategoricalVectors

Keywords

features: Either:

A Vector of AbstractFeatureClass type features; or

A string where "l" = linear and categorical, "q" = quadratic, "p" = product, "t" = threshold, "h" = hinge; or

Nothing, in which case the default features based on the number of presences are used

regularization_multiplier: A constant to adjust regularization, where a higher regularization_multiplier results in a higher penalization for features

regularization_function: A function to compute a regularization for each feature. A default regularization_function is built in.

addsamplestobackground: A boolean, where true adds the background samples to the predictors. Defaults to true.

n_knots: the number of knots used for Threshold and Hinge features. Defaults to 50. Ignored if there are neither Threshold nor Hinge features

weight_factor: A Float64 to adjust the weight of the background samples. Defaults to 100.0.

backend: Either LassoBackend() or GLMNetBackend(), to use Lasso.jl or GLMNet.jl fit the model.

Lasso.jl is written in pure julia, but can be slower with large model matrices (e.g. when hinge is enabled). Defaults to LassoBackend.

kw...: Further arguments to be passed to Lasso.fit or GLMNet.glmnet

Returns

model: A model of type MaxnetModel